update maps a bit and started new cutScene
map were just modified a little to add shit!
the route to prison map has a shortcut so we can go to end faster!
the prison map has tons of enemies!
the first map now spawns the player near the saloon and makes him walk a little! this is going to be a cutscene... but we will see!